move more functions to block_reader
