+++ src/query/storages/fuse/src/io/read/block_reader.rs
@@ -16,18 +16,23 @@ use std::collections::HashMap;
 use std::ops::Range;
 use std::time::Instant;
 
+use common_arrow::arrow::datatypes::Field;
 use common_arrow::parquet::metadata::SchemaDescriptor;
 use common_base::rangemap::RangeMerger;
 use common_base::runtime::UnlimitedFuture;
+use common_catalog::plan::PartInfoPtr;
 use common_catalog::plan::Projection;
 use common_datavalues::DataSchemaRef;
 use common_exception::ErrorCode;
 use common_exception::Result;
+use common_storage::ColumnLeaf;
 use common_storage::ColumnLeaves;
+use common_storages_table_meta::meta::ColumnMeta;
 use futures::future::try_join_all;
 use opendal::Object;
 use opendal::Operator;
 
+use crate::fuse_part::FusePartInfo;
 use crate::io::read::ReadSettings;
 use crate::metrics::*;
 
@@ -102,6 +107,10 @@ where Self: 'static
 }
 
 impl BlockReader {
+    pub fn support_blocking_api(&self) -> bool {
+        self.operator.metadata().can_blocking()
+    }
+
     /// This is an optimized for data read, works like the Linux kernel io-scheduler IO merging.
     /// If the distance between two IO request ranges to be read is less than storage_io_min_bytes_for_seek(Default is 48Bytes),
     /// will read the range that contains both ranges, thus avoiding extra seek.
@@ -226,6 +235,73 @@ impl BlockReader {
         }
 
         Ok(read_res)
+    }
+
+    pub async fn read_columns_data_by_merge_io(
+        &self,
+        settings: &ReadSettings,
+        location: &str,
+        columns_meta: &HashMap<usize, ColumnMeta>,
+    ) -> Result<MergeIOReadResult> {
+        // Perf
+        {
+            metrics_inc_remote_io_read_parts(1);
+        }
+
+        let columns = self.projection.project_column_leaves(&self.column_leaves)?;
+        let indices = Self::build_projection_indices(&columns);
+
+        let mut ranges = vec![];
+        for index in indices.keys() {
+            let column_meta = &columns_meta[index];
+            ranges.push((
+                *index,
+                column_meta.offset..(column_meta.offset + column_meta.len),
+            ));
+
+            // Perf
+            {
+                metrics_inc_remote_io_seeks(1);
+                metrics_inc_remote_io_read_bytes(column_meta.len);
+            }
+        }
+
+        let object = self.operator.object(location);
+
+        Self::merge_io_read(settings, object, ranges).await
+    }
+
+    pub fn sync_read_columns_data_by_merge_io(
+        &self,
+        settings: &ReadSettings,
+        part: PartInfoPtr,
+    ) -> Result<MergeIOReadResult> {
+        let part = FusePartInfo::from_part(&part)?;
+        let columns = self.projection.project_column_leaves(&self.column_leaves)?;
+        let indices = Self::build_projection_indices(&columns);
+
+        let mut ranges = vec![];
+        for index in indices.keys() {
+            let column_meta = &part.columns_meta[index];
+            ranges.push((
+                *index,
+                column_meta.offset..(column_meta.offset + column_meta.len),
+            ));
+        }
+
+        let object = self.operator.object(&part.location);
+        Self::sync_merge_io_read(settings, object, ranges)
+    }
+
+    // Build non duplicate leaf_ids to avoid repeated read column from parquet
+    pub(crate) fn build_projection_indices(columns: &Vec<&ColumnLeaf>) -> HashMap<usize, Field> {
+        let mut indices = HashMap::with_capacity(columns.len());
+        for column in columns {
+            for index in &column.leaf_ids {
+                indices.insert(*index, column.field.clone());
+            }
+        }
+        indices
     }
 
     #[inline]
